examples/drive-thru/requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
annoy==1.17.3
anyio==4.7.0
async-timeout==4.0.3
attrs==24.3.0
cartesia==1.0.14
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepgram-sdk==3.7.7
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
flatbuffers==24.3.25
frozenlist==1.5.0
future==1.0.0
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
humanfriendly==10.0
idna==3.10
iniconfig==2.0.0
iterators==0.2.0
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-openai==0.2.12
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.3
langsmith==0.2.3
llvmlite==0.43.0
loguru==0.7.3
Markdown==3.7
marshmallow==3.23.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
numba==0.60.0
numpy==1.26.4
onnxruntime==1.20.1
openai==1.57.4
orjson==3.10.12
packaging==24.2
pillow==10.4.0
pinecone==5.4.2
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pipecat-ai @ git+https://github.com/pipecat-ai/pipecat.git@09a611d44ba09c8b875c95fc6cdfcc98d3b7bee3
pluggy==1.5.0
propcache==0.2.1
protobuf==5.29.1
pydantic==2.8.2
pydantic-settings==2.7.0
pydantic_core==2.20.1
pyloudnorm==0.1.1
pytest==8.3.4
pytest-asyncio==0.25.0
python-dateutil==2.9.0.post0
python-deepcompare==1.0.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
resampy==0.4.3
scipy==1.14.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.3
tenacity==9.0.0
tiktoken==0.8.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.34.0
websockets==13.1
yarl==1.18.3
